requirements.txt
# Dépendances pour le projet Smart Bus IoT

# GPS Neo-6M
pyserial>=3.5
pynmea2>=1.19.0

# DHT22 (Température/Humidité)
Adafruit-Python-DHT>=1.4.0

# MPU9250 (IMU)
mpu9250-jmdev>=1.0.0

# GPIO pour Raspberry Pi
RPi.GPIO>=0.7.1

# LCD I2C
RPLCD>=0.9.0

# Utilitaires
python-dotenv>=1.0.0





